requirements.txt
Flask==3.0.3
Jinja2==3.1.4
python-nmap==0.7.1
scapy==2.5.0
SQLAlchemy==2.0.32
psutil==6.0.0
requests==2.32.3
